+++ test_support/HeaderHygiene.cpp
@@ -67,6 +67,8 @@
 #include <felitronics/lineareq/MagnitudeCurve.h>
 #include <felitronics/lineareq/MixedPhaseFir.h>
 #include <felitronics/lineareq/NaturalPhaseEq.h>
+#include <felitronics/mastering/MasteringChain.h>
+#include <felitronics/mastering/OfflineRenderer.h>
 #include <felitronics/measurement/CaptureGate.h>
 #include <felitronics/measurement/Convolve.h>
 #include <felitronics/measurement/Deconvolve.h>